mainpro2/requirements.txt
streamlit==1.27.0
numpy==1.24.3
opencv-python==4.8.0.76
pandas==2.0.3
matplotlib==3.7.2
pillow==9.5.0